backend/requirements.txt
# A2A Protocol & Core
python-a2a>=0.5.0
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic[email]>=2.5.0
pydantic-settings>=2.0.0
email-validator>=2.0.0
httpx>=0.25.0
sse-starlette>=1.8.0

# LLM Providers
openai>=1.3.0
anthropic>=0.18.0
google-generativeai>=0.3.0

# Async support
aiohttp>=3.9.0
asyncio-throttle>=1.0.0

# Utilities
python-dotenv>=1.0.0
loguru>=0.7.0
tenacity>=8.2.0

# CORS
python-multipart>=0.0.6

# Database
asyncpg>=0.29.0
sqlalchemy[asyncio]>=2.0.0

# Cache
redis>=5.0.0

# Authentication
passlib[bcrypt]>=1.7.4
python-jose[cryptography]>=3.3.0
bcrypt>=4.0.0
cryptography>=41.0.0

# OpenTelemetry (Programmatic instrumentation)
opentelemetry-api>=1.27.0
opentelemetry-sdk>=1.27.0
opentelemetry-exporter-otlp>=1.27.0
opentelemetry-instrumentation-fastapi>=0.48b0
opentelemetry-instrumentation-httpx>=0.48b0

